Add a User Interface and refactor the JS code
The game will be played with a User interface.
There will be three buttons to select an item,
a counter for the player and the computer
and an info field with instructions
and an announcement who's the winner of the round
The game will end after one player has five points.